+++ src/modules/pipeline/step3-script/frontend/Container.tsx
@@ -9,8 +9,8 @@ import { usePipelineStore } from "../../../../renderer/src/store/usePipelineStor
 import { useProjectStore } from "../../../editor/stores/useProjectStore";
 import { API } from "../../../../renderer/src/api";
 import { mapPipelineResultToState } from "../../../../renderer/src/pages/editor/hooks/usePipelineResultMapper";
-import { buildMappers } from "../../../../renderer/src/pages/editor/hooks/usePipelineOrchestrator";
-import { STEP_SEQUENCES } from "../../../../renderer/src/pages/editor/utils/pipelineConstants";
+import { buildMappers } from "../../../../editor/shell/frontend/hooks/usePipelineOrchestrator";
+import { STEP_SEQUENCES } from "../../../../editor/shell/utils/pipelineConstants";
 import { diffParagraphs, applyDiffUpdate } from "../../../../renderer/src/pages/editor/utils/scriptDiffTree";
 import { StepScriptGenerationView } from "./View";