+++ src/modules/pipeline/step5-match/frontend/Container.tsx
@@ -11,8 +11,8 @@ import { usePipelineStore } from "../../../../renderer/src/store/usePipelineStor
 import { useProjectStore } from "../../../editor/stores/useProjectStore";
 import { API } from "../../../../renderer/src/api";
 import { mapPipelineResultToState } from "../../../../renderer/src/pages/editor/hooks/usePipelineResultMapper";
-import { buildMappers } from "../../../../renderer/src/pages/editor/hooks/usePipelineOrchestrator";
-import { STEP_SEQUENCES } from "../../../../renderer/src/pages/editor/utils/pipelineConstants";
+import { buildMappers } from "../../../../editor/shell/frontend/hooks/usePipelineOrchestrator";
+import { STEP_SEQUENCES } from "../../../../editor/shell/utils/pipelineConstants";
 import { StepShotMatchingView } from "./View";
 
 export const StepShotMatching: React.FC = () => {